# docs/swagger.yaml
basePath: /
definitions:
...
  dto.GetUserResponse:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
...
  dto.NewUserRequest:
    properties:
      age:
        example: 21
        type: integer
      email:
        example: monday.day@email.com
        type: string
      password:
        example: secret
        type: string
      username:
        example: monday
        type: string
    type: object
...
  dto.UserLoginRequest:
    properties:
      email:
        example: monday.day@email.com
        type: string
      password:
        example: secret
        type: string
    type: object
  dto.UserUpdateRequest:
    properties:
      email:
        example: monday.day@weeekly.com
        type: string
      username:
        example: monday
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Final Project 1
  title: MyGram App
  version: "1.0"
paths:
...
  /users:
    delete:
      consumes:
      - application/json
      description: Delete Current User
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetUserResponse'
      security:
      - BearerAuth: []
      summary: Delete Current User
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: User update
      parameters:
      - description: body request for user update
        in: body
        name: dto.UserUpdateRequest
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetUserResponse'
      security:
      - BearerAuth: []
      summary: User update
      tags:
      - Users
  /users/login:
    post:
      consumes:
      - application/json
      description: User login
      parameters:
      - description: body request for user login
        in: body
        name: dto.UserLoginRequest
        required: true
        schema:
          $ref: '#/definitions/dto.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetUserResponse'
      summary: User login
      tags:
      - Users
  /users/register:
    post:
      consumes:
      - application/json
      description: User register
      parameters:
      - description: body request for user register
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/dto.NewUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.GetUserResponse'
      summary: User register
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and token you got from the User
      Login api.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"